Fix an issue in bt_split_child; remove OST (b-tree)
I am giving up on the goal of creating an OST from a B-Tree. This would
require modifying the B-tree structure by adding a parent pointer to each node.
Instead, I suggest using a Red-Black Tree for this purpose. B-Trees are better
suited for tasks such as maintaining ordered sets or dictionaries.